Relicence to MPL
This is intended to be retroactive, any previous version of the code in this repository is also now available under the MPL as well as the GPL, at your option.

From this commit, the code is solely available under the MPL, not the GPL.

All copyright in this repository is owned by me (jamie@quigley.xyz), so re-licencing is possible.